Update
trying to fix non law stat on spreadsheet

+++ WorkGui/Student.cs
@@ -51,9 +51,9 @@ namespace classesForLibraryExcel
                 public int getCredsInProgress() { return credsInProgress; }
                 public void setCredsInProgress(int credsInProgress1) { credsInProgress = credsInProgress1; }
                 public void addCredsInProgress(int addCred) { credsInProgress += addCred; }
-        private int totCred = 0; //total credits
-                public int getTotCred() { return totCred; }
-                public void setTotCred(int totCred1) { totCred = totCred1; }
+        private double totCred = 0.0; //total credits
+                public double getTotCred() { return totCred; }
+                public void setTotCred(double totCred1) { totCred = totCred1; }
         private double enrollUnits = 0.0; //enrollment units
                 public void setEnrollUnits(double units) { enrollUnits = units; }
                 public double getEnrollUnits() { return enrollUnits; }
